Refactor MXFP8 cute SM120 to grouped_mm/ scope (PR scope shrink)
After PM feedback, this PR's exposed surface shrinks to only the MoE
zero_padding entry + its dedicated quantization helper. The other 4
GEMM variants (dense / batched / contiguous / masked) are deferred to
future PRs; their kernel-level infrastructure is preserved (the runner
class still declares all 5 methods) but no FFI bindings, Python API,
or tests reference them.

Structural moves
- include/flashinfer/gemm/{mxfp8_gemm_cute_sm120.h, sm120_blockscaled/}
    -> include/flashinfer/grouped_mm/
- csrc/{mxfp8_gemm_cute_sm120.cu, group_gemm_mxfp8_cute_sm120.cu,
       group_gemm_mxfp8_cute_sm120_jit_binding.cu}
    -> csrc/fused_moe/cutlass_backend/cute_sm120/
- flashinfer/jit/gemm/core.py: gen_gemm_sm120_module_cute_mxfp8
    -> flashinfer/jit/grouped_mm/core.py: gen_grouped_mm_sm120_module_cute_mxfp8
- flashinfer/gemm/gemm_base.py (5 entries + 3 helpers + 1 accessor)
    -> flashinfer/grouped_mm/core.py (zero_padding + quantize + helpers)
- tests/gemm/test_groupwise_scaled_gemm_mxfp8.py (116 cells, 5 tests)
    -> tests/grouped_mm/test_grouped_mm_mxfp8_zero_padding.py (16 cells, 1 test)
- docs/api/gemm.rst MXFP8 section
    -> docs/api/grouped_mm.rst (new)

Strip
- Dense+BMM csrc binding files deleted
  (csrc/{gemm_mxfp8_cute_sm120,gemm_mxfp8_cute_sm120_jit_binding}.cu)
- Grouped FFI binding stripped from 4 entries to 2
  (zero_padding + quantize_mxfp8_for_zero_padding kept)
- Python API: 4 dropped entries (dense / BMM / contiguous / masked);
  zero_padding entry renamed for namespace alignment
  group_gemm_mxfp8_nt_groupwise_zero_padding
    -> grouped_mm_mxfp8_nt_groupwise_zero_padding

Verify
- Cold JIT cache rebuild + 16/16 pytest PASS (156s)
