+++ docs/index.rst
@@ -34,6 +34,7 @@ FlashInfer is a library and kernel generator for Large Language Models that prov
 
    api/attention
    api/gemm
+   api/grouped_mm
    api/fused_moe
    api/cascade
    api/comm
