+++ flashinfer/aot.py
@@ -69,6 +69,7 @@ from .jit.fused_moe import (
 )
 from .jit.bgmv_moe import gen_bgmv_moe_module
 from .jit.gdn import gen_gdn_prefill_sm90_module
+from .jit.grouped_mm import gen_grouped_mm_sm120_module_cute_mxfp8
 from .jit.gemm import (
     gen_fp8_blockscale_gemm_sm90_module,
     gen_gemm_module,
@@ -78,7 +79,6 @@ from .jit.gemm import (
     gen_gemm_sm100_module_cutlass_fp8,
     gen_gemm_sm100_module_cutlass_mxfp8,
     gen_gemm_sm120_module,
-    gen_gemm_sm120_module_cute_mxfp8,
     gen_gemm_sm120_module_cutlass_fp4,
     gen_mm_bf16_cublaslt_module,
     gen_gemm_sm120_module_cutlass_mxfp8,
@@ -546,7 +546,7 @@ def gen_all_modules(
             # compiles for all SM12x targets.
             jit_specs.append(gen_cutlass_fused_moe_sm120_module())
             jit_specs.append(gen_gemm_sm120_module())
-            jit_specs.append(gen_gemm_sm120_module_cute_mxfp8())
+            jit_specs.append(gen_grouped_mm_sm120_module_cute_mxfp8())
             jit_specs.append(gen_gemm_sm120_module_cutlass_fp4())
             jit_specs.append(gen_gemm_sm120_module_cutlass_mxfp8())
             jit_specs.append(gen_trtllm_fmha_v2_sm120_module())
