+++ flashinfer/gemm/gemm_base.py
@@ -64,7 +64,6 @@ from ..jit.gemm import gen_gemm_sm90_module
 from ..jit.gemm import gen_gemm_module
 from ..jit.gemm import gen_gemm_sm100_module
 from ..jit.gemm import gen_gemm_sm120_module
-from ..jit.gemm import gen_gemm_sm120_module_cute_mxfp8
 from ..jit.gemm import gen_gemm_sm120_module_cutlass_fp4
 from ..jit.gemm import gen_gemm_sm100_module_cutlass_fp4
 from ..jit.gemm import gen_gemm_sm103_module_cutlass_fp4
@@ -7134,641 +7133,6 @@ def group_gemm_fp8_nt_groupwise(
             scale_major_mode,
             mma_sm,
         )
-    return out
-
-
-@functools.cache
-def get_gemm_sm120_module_cute_mxfp8():
-    """MXFP8 group GEMM module accessor for SM120."""
-    return gen_gemm_sm120_module_cute_mxfp8().build_and_load()
-
-
-def _check_m_indptr(m_indptr: torch.Tensor, token_num: int) -> None:
-    """Validate ``m_indptr`` semantic consistency before launching the kernel.
-
-    Checks: 1-D int32, first element 0, monotonic non-decreasing, last element
-    equals ``token_num`` (the row count of the packed A tensor).
-    """
-    if m_indptr.dtype != torch.int32:
-        raise ValueError(f"m_indptr must be int32; got {m_indptr.dtype}")
-    if m_indptr.dim() != 1:
-        raise ValueError(f"m_indptr must be 1-D; got {m_indptr.dim()}D")
-    if m_indptr.numel() < 2:
-        raise ValueError(
-            f"m_indptr must have at least 2 elements (num_experts >= 1); "
-            f"got numel={m_indptr.numel()}"
-        )
-    first = int(m_indptr[0].item())
-    last = int(m_indptr[-1].item())
-    if first != 0:
-        raise ValueError(f"m_indptr[0] must be 0; got {first}")
-    if last != token_num:
-        raise ValueError(f"m_indptr[-1] must equal token_num={token_num}; got {last}")
-    if not bool((m_indptr[1:] >= m_indptr[:-1]).all().item()):
-        raise ValueError("m_indptr must be non-decreasing")
-
-
-def _check_scale_granularity_mnk(scale_granularity_mnk: Tuple[int, int, int]) -> None:
-    """Validate the per-token UE8M0 ``scale_granularity_mnk`` contract shared by all
-    MXFP8 cute SM120 GEMM entries. Accepts ``(1, 1, 32)`` or ``(1, 1, 128)``.
-    """
-    if len(scale_granularity_mnk) != 3:
-        raise ValueError(
-            f"scale_granularity_mnk must be a 3-tuple (m_gran, n_gran, k_gran); "
-            f"got length {len(scale_granularity_mnk)}"
-        )
-    if scale_granularity_mnk[0] != 1:
-        raise ValueError(
-            f"scale_granularity_mnk[0] (m_gran) must be 1; got {scale_granularity_mnk[0]}"
-        )
-    if scale_granularity_mnk[1] != 1:
-        raise ValueError(
-            f"scale_granularity_mnk[1] (n_gran) must be 1 (kernel only exposes granK "
-            f"along K; 2D block B-scale must be broadcast to per-token caller-side); "
-            f"got {scale_granularity_mnk[1]}"
-        )
-    if scale_granularity_mnk[2] not in (32, 128):
-        raise ValueError(
-            f"scale_granularity_mnk[2] (k_gran) must be 32 or 128; got {scale_granularity_mnk[2]}"
-        )
-
-
-def _check_scale_major_mode_mxfp8(scale_major_mode: str) -> None:
-    """Validate ``scale_major_mode`` for the MXFP8 cute SM120 GEMM entries.
-
-    The kernel only consumes per-token INT32-packed UE8M0 scales in MN-major
-    TMA-aligned layout. Future K-major support would require kernel changes;
-    until then any value other than ``"MN"`` is rejected.
-    """
-    if scale_major_mode != "MN":
-        raise ValueError(
-            f'scale_major_mode must be "MN" (kernel currently only supports the '
-            f'per-token MN-major TMA-aligned UE8M0 scale layout); got "{scale_major_mode}"'
-        )
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def quantize_mxfp8_for_zero_padding(
-    input: torch.Tensor,
-    m_indptr: torch.Tensor,
-    gran_k: int = 128,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    r"""Per-token MXFP8 quantization for ``group_gemm_mxfp8_nt_groupwise`` zero_padding mode.
-
-    Produces a token-packed FP8 output and a 4-row per-expert padded int32-packed UE8M0
-    scale tensor matching the input layout expected by ``group_gemm_mxfp8_nt_groupwise``
-    with ``mode="zero_padding"``.
-
-    Parameters
-    ----------
-    input: torch.Tensor
-        Input tensor shape ``(token_num, k)``, data type is ``torch.bfloat16``.
-
-    m_indptr: torch.Tensor
-        The indptr of the segment lengths, shape ``(num_experts + 1,)``, data type is
-        ``torch.int32``. ``m_indptr[0] = 0``, ``m_indptr[num_experts] = token_num``.
-
-    gran_k: int
-        UE8M0 K-axis block granularity. Must be ``32`` or ``128``.
-
-    Returns
-    -------
-    out_fp8: torch.Tensor
-        Token-packed FP8 output, shape ``(token_num, k)``, data type is
-        ``torch.float8_e4m3fn``.
-
-    out_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor (4 UE8M0 scales packed per int32), shape
-        ``(m_padded, k_align)`` where ``m_padded = (token_num + num_experts * 3) // 4 * 4``
-        and ``k_align = (k + 4 * gran_k - 1) // (4 * gran_k)``. Data type is ``torch.int32``.
-        Underlying storage is ``(k_align, m_padded)`` contiguous; the returned view is a
-        ``.transpose(0, 1)`` that aligns with the ``group_gemm_mxfp8_nt_groupwise`` caller
-        convention.
-    """
-    if gran_k not in (32, 128):
-        raise ValueError(f"gran_k must be 32 or 128; got {gran_k}")
-    if input.dtype != torch.bfloat16:
-        raise ValueError(f"input must be bfloat16; got {input.dtype}")
-    if input.dim() != 2:
-        raise ValueError(f"input must be 2D; got {input.dim()}D")
-
-    token_num, k = input.shape
-    _check_m_indptr(m_indptr, token_num=token_num)
-    num_experts = m_indptr.shape[0] - 1
-    if k % 16 != 0:
-        raise ValueError(f"k must be multiple of 16; got k={k}")
-
-    pack_nk = gran_k * 4
-    m_padded = (token_num + num_experts * 3) // 4 * 4
-    k_align = (k + pack_nk - 1) // pack_nk
-
-    out_fp8 = torch.empty(
-        (token_num, k), dtype=torch.float8_e4m3fn, device=input.device
-    )
-    out_scale_raw = torch.zeros(
-        (k_align, m_padded), dtype=torch.int32, device=input.device
-    )
-
-    get_gemm_sm120_module_cute_mxfp8().quantize_mxfp8_for_zero_padding(
-        input, m_indptr, out_fp8, out_scale_raw, gran_k
-    )
-
-    out_scale = out_scale_raw.transpose(0, 1)
-    return out_fp8, out_scale
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def group_gemm_mxfp8_nt_groupwise_zero_padding(
-    a: torch.Tensor,
-    b: torch.Tensor,
-    a_scale: torch.Tensor,
-    b_scale: torch.Tensor,
-    m_indptr: torch.Tensor,
-    scale_granularity_mnk: Tuple[int, int, int] = (1, 1, 128),
-    scale_major_mode: Literal["MN"] = "MN",
-    backend: Literal["cute"] = "cute",
-    out: Optional[torch.Tensor] = None,
-    out_dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Perform group GEMM with MXFP8 inputs in zero-padding mode using groupwise UE8M0
-    scaling. Currently only supported on NVIDIA RTX PRO 6000 Blackwell (SM120) architecture.
-
-    Zero-padding mode accepts token-packed input ``a`` (no per-expert pre-padding along M)
-    with 4-row per-expert padding on the scale tensor ``a_scale``. The group descriptor is
-    a CSR cumsum ``m_indptr``. This mode is optimized for decoding with small per-expert M
-    (down to ``m_per_expert = 1``) where DeepGEMM-style contiguous padding would waste
-    memory and compute.
-
-    Parameters
-    ----------
-    a: torch.Tensor
-        Row-major input tensor shape ``(cum_m, k)``, data type is ``torch.float8_e4m3fn``.
-        Token-packed across experts; ``cum_m`` is the cumulative sum of segment lengths.
-
-    b: torch.Tensor
-        Column-major input tensor shape ``(num_experts, n, k)``, data type is
-        ``torch.float8_e4m3fn``.
-
-    a_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``a`` (4 UE8M0 scales packed per int32), shape
-        ``(m_padded, k_align)`` where ``m_padded = (cum_m + num_experts * 3) // 4 * 4`` and
-        ``k_align = (k + 4 * k_granularity - 1) // (4 * k_granularity)``. Data type is
-        ``torch.int32``.
-
-    b_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``b`` in per-token layout, shape
-        ``(num_experts, n, k_align)``. Data type is ``torch.int32``. See Notes for the
-        per-token layout requirement.
-
-    m_indptr: torch.Tensor
-        The indptr of the segment lengths, shape ``(num_experts + 1,)``, data type is
-        ``torch.int32``. ``m_indptr[0] = 0``, ``m_indptr[num_experts] = cum_m``.
-
-    scale_granularity_mnk: Tuple[int, int, int]
-        The granularity of the scale tensor, ``(m_granularity, n_granularity, k_granularity)``.
-        Accepted values: ``(1, 1, 128)`` (DeepGEMM-style production, default) or
-        ``(1, 1, 32)`` (OCP MXFP8). ``m_granularity`` and ``n_granularity`` must both be
-        ``1`` (per-token scaling along M and N); ``k_granularity`` must be ``32`` or ``128``.
-        Anything else raises ``ValueError``.
-
-    backend: Literal["cute"]
-        Backend selector. Currently only ``"cute"`` is implemented.
-
-    out: Optional[torch.Tensor]
-        The output tensor, shape ``(cum_m, n)``. If not specified, an output tensor will be
-        created.
-
-    out_dtype: Optional[torch.dtype]
-        The data type of the output tensor. Currently only ``torch.bfloat16`` is supported.
-
-    Returns
-    -------
-    out: torch.Tensor
-        The output tensor, shape ``(cum_m, n)``.
-
-    Notes
-    -----
-    - MXFP8 uses UE8M0 scales over K-axis blocks of size 32 (OCP spec) or 128
-      (DeepGEMM convention). Both ``a_scale`` and ``b_scale`` must be provided in per-token
-      layout: one UE8M0 scale per row along M (for ``a``) or N (for ``b``), packed 4 scales
-      per int32 along the K-axis blocks.
-    - If a caller starts from a 2D ``(k_granularity, k_granularity)`` block-quantized
-      ``b_scale``, it must be broadcast to per-token shape ``(num_experts, n, k_align)``
-      before invoking this function (one scale per N-row).
-    """
-    _check_scale_granularity_mnk(scale_granularity_mnk)
-    _check_scale_major_mode_mxfp8(scale_major_mode)
-    _check_m_indptr(m_indptr, token_num=a.shape[0])
-    if backend != "cute":
-        raise NotImplementedError(
-            f'Only backend="cute" is currently implemented; got backend="{backend}"'
-        )
-
-    if out_dtype is None:
-        out_dtype = out.dtype if out is not None else torch.bfloat16
-    if out_dtype != torch.bfloat16:
-        raise NotImplementedError(
-            f"Only out_dtype=torch.bfloat16 is supported; got {out_dtype}"
-        )
-
-    n = b.shape[1]
-    if out is None:
-        out = torch.empty((a.shape[0], n), dtype=out_dtype, device=a.device)
-
-    get_gemm_sm120_module_cute_mxfp8().group_gemm_mxfp8_nt_groupwise_zero_padding(
-        a,
-        b,
-        a_scale,
-        b_scale,
-        m_indptr,
-        out,
-        scale_major_mode,
-        scale_granularity_mnk[0],
-        scale_granularity_mnk[1],
-        scale_granularity_mnk[2],
-    )
-    return out
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def group_gemm_mxfp8_nt_groupwise(
-    a: torch.Tensor,
-    b: torch.Tensor,
-    a_scale: torch.Tensor,
-    b_scale: torch.Tensor,
-    m_indices: torch.Tensor,
-    scale_granularity_mnk: Tuple[int, int, int] = (1, 1, 128),
-    scale_major_mode: Literal["MN"] = "MN",
-    use_psum_layout: bool = False,
-    backend: Literal["cute"] = "cute",
-    out: Optional[torch.Tensor] = None,
-    out_dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Perform group GEMM with MXFP8 inputs using groupwise UE8M0 scaling, with the
-    DeepGEMM-style contiguous group layout. Signature aligned with
-    ``flashinfer.deep_gemm.m_grouped_fp8_gemm_nt_contiguous``. Currently only supported on
-    NVIDIA RTX PRO 6000 Blackwell (SM120) architecture.
-
-    Parameters
-    ----------
-    a: torch.Tensor
-        Row-major input tensor shape ``(m, k)``, data type is ``torch.float8_e4m3fn``.
-
-    b: torch.Tensor
-        Column-major input tensor shape ``(num_experts, n, k)``, data type is
-        ``torch.float8_e4m3fn``.
-
-    a_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``a`` in per-token layout, shape
-        ``(m, k_align)``. Data type is ``torch.int32``.
-
-    b_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``b`` in per-token layout, shape
-        ``(num_experts, n, k_align)``. Data type is ``torch.int32``.
-
-    m_indices: torch.Tensor
-        Group descriptor, data type is ``torch.int32``. Shape depends on ``use_psum_layout``:
-
-        - ``use_psum_layout=False`` (contiguous): shape ``(m,)`` per-token group assignment;
-          ``m_indices[i] = j`` means row ``i`` in ``a`` is multiplied with group ``j`` in
-          ``b``. Follows the DeepGEMM convention.
-        - ``use_psum_layout=True``: shape ``(num_experts,)`` cumulative sum of per-group
-          aligned M (no leading zero).
-
-    scale_granularity_mnk: Tuple[int, int, int]
-        The granularity of the scale tensor, ``(m_granularity, n_granularity, k_granularity)``.
-        Accepted values: ``(1, 1, 128)`` (default) or ``(1, 1, 32)``. ``m_granularity``
-        and ``n_granularity`` must both be ``1`` (per-token scaling); ``k_granularity`` must
-        be ``32`` or ``128``.
-
-    use_psum_layout: bool
-        If ``True``, dispatch to the psum-layout kernel variant which accepts pre-padded
-        per-expert M segments. Default ``False`` (contiguous layout).
-
-    out: Optional[torch.Tensor]
-        The output tensor, shape ``(m, n)``. If not specified, an output tensor will be
-        created.
-
-    out_dtype: Optional[torch.dtype]
-        The data type of the output tensor. Currently only ``torch.bfloat16`` is supported.
-
-    Returns
-    -------
-    out: torch.Tensor
-        The output tensor, shape ``(m, n)``.
-    """
-    _check_scale_granularity_mnk(scale_granularity_mnk)
-    _check_scale_major_mode_mxfp8(scale_major_mode)
-    if backend != "cute":
-        raise NotImplementedError(
-            f'Only backend="cute" is currently implemented; got backend="{backend}"'
-        )
-
-    if out_dtype is None:
-        out_dtype = out.dtype if out is not None else torch.bfloat16
-    if out_dtype != torch.bfloat16:
-        raise NotImplementedError(
-            f"Only out_dtype=torch.bfloat16 is supported; got {out_dtype}"
-        )
-
-    m = a.shape[0]
-    n = b.shape[1]
-    if out is None:
-        out = torch.empty((m, n), dtype=out_dtype, device=a.device)
-
-    get_gemm_sm120_module_cute_mxfp8().group_gemm_mxfp8_nt_groupwise(
-        a,
-        b,
-        a_scale,
-        b_scale,
-        m_indices,
-        out,
-        use_psum_layout,
-        scale_granularity_mnk[0],
-        scale_granularity_mnk[1],
-        scale_granularity_mnk[2],
-    )
-    return out
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def group_gemm_mxfp8_nt_groupwise_masked(
-    a: torch.Tensor,
-    b: torch.Tensor,
-    a_scale: torch.Tensor,
-    b_scale: torch.Tensor,
-    masked_m: torch.Tensor,
-    scale_granularity_mnk: Tuple[int, int, int] = (1, 1, 128),
-    scale_major_mode: Literal["MN"] = "MN",
-    backend: Literal["cute"] = "cute",
-    out: Optional[torch.Tensor] = None,
-    out_dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Perform masked grouped GEMM with MXFP8 inputs using groupwise UE8M0 scaling.
-    Signature aligned with ``flashinfer.deep_gemm.m_grouped_fp8_gemm_nt_masked``.
-    Currently only supported on NVIDIA RTX PRO 6000 Blackwell (SM120) architecture.
-
-    Each expert's A slice ``a[i]`` is multiplied with ``b[i]``; only the first
-    ``masked_m[i]`` rows of ``a[i]`` are considered valid (the rest are ignored).
-
-    Parameters
-    ----------
-    a: torch.Tensor
-        Per-expert padded input tensor shape ``(num_experts, max_m, k)``, data type
-        is ``torch.float8_e4m3fn``.
-
-    b: torch.Tensor
-        Per-expert weight tensor shape ``(num_experts, n, k)``, data type is
-        ``torch.float8_e4m3fn``.
-
-    a_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``a``, shape
-        ``(num_experts, max_m, k_align)`` per-token layout. Data type ``torch.int32``.
-
-    b_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``b``, shape
-        ``(num_experts, n, k_align)`` per-token layout. Data type ``torch.int32``.
-
-    masked_m: torch.Tensor
-        Per-group valid row count, shape ``(num_experts,)``, data type
-        ``torch.int32``. ``masked_m[i] = j`` means only the first ``j`` rows of
-        ``a[i]`` are computed; the rest are ignored. Mirrors DG / 6KD canonical
-        naming.
-
-    scale_granularity_mnk: Tuple[int, int, int]
-        Accepted values: ``(1, 1, 128)`` (default) or ``(1, 1, 32)``.
-
-    out: Optional[torch.Tensor]
-        Output tensor, shape ``(num_experts, max_m, n)``. If not specified, created.
-
-    out_dtype: Optional[torch.dtype]
-        Output dtype. Currently only ``torch.bfloat16`` is supported.
-
-    Returns
-    -------
-    out: torch.Tensor
-        Output tensor, shape ``(num_experts, max_m, n)``. Only the first
-        ``masked_m[i]`` rows per expert are valid.
-    """
-    _check_scale_granularity_mnk(scale_granularity_mnk)
-    _check_scale_major_mode_mxfp8(scale_major_mode)
-    if backend != "cute":
-        raise NotImplementedError(
-            f'Only backend="cute" is currently implemented; got backend="{backend}"'
-        )
-
-    if out_dtype is None:
-        out_dtype = out.dtype if out is not None else torch.bfloat16
-    if out_dtype != torch.bfloat16:
-        raise NotImplementedError(
-            f"Only out_dtype=torch.bfloat16 is supported; got {out_dtype}"
-        )
-
-    num_experts, max_m, _ = a.shape
-    n = b.shape[1]
-    if out is None:
-        out = torch.empty((num_experts, max_m, n), dtype=out_dtype, device=a.device)
-
-    get_gemm_sm120_module_cute_mxfp8().group_gemm_mxfp8_nt_groupwise_masked(
-        a,
-        b,
-        a_scale,
-        b_scale,
-        masked_m,
-        out,
-        scale_granularity_mnk[0],
-        scale_granularity_mnk[1],
-        scale_granularity_mnk[2],
-    )
-    return out
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def gemm_mxfp8_nt_groupwise(
-    a: torch.Tensor,
-    b: torch.Tensor,
-    a_scale: torch.Tensor,
-    b_scale: torch.Tensor,
-    scale_granularity_mnk: Tuple[int, int, int] = (1, 1, 128),
-    scale_major_mode: Literal["MN"] = "MN",
-    backend: Literal["cute"] = "cute",
-    out: Optional[torch.Tensor] = None,
-    out_dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Perform single MXFP8 matmul with groupwise UE8M0 scaling (dense Normal entry).
-    Currently only supported on NVIDIA RTX PRO 6000 Blackwell (SM120) architecture.
-
-    Parameters
-    ----------
-    a: torch.Tensor
-        Row-major input tensor shape ``(m, k)``, data type is ``torch.float8_e4m3fn``.
-
-    b: torch.Tensor
-        Column-major weight tensor shape ``(n, k)``, data type is ``torch.float8_e4m3fn``.
-
-    a_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``a``, shape ``(m, k_align)`` per-token layout.
-        Data type ``torch.int32``.
-
-    b_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``b``, shape ``(n, k_align)`` per-token layout.
-        Data type ``torch.int32``.
-
-    scale_granularity_mnk: Tuple[int, int, int]
-        Accepted values: ``(1, 1, 128)`` (default) or ``(1, 1, 32)``. Per-row contract
-        (cute backend) — distinct from DG legacy ``(1, 128, 128)`` 2D-block contract.
-
-    out: Optional[torch.Tensor]
-        Output tensor, shape ``(m, n)``. If not specified, created internally.
-
-    out_dtype: Optional[torch.dtype]
-        Output dtype. Currently only ``torch.bfloat16`` is supported.
-
-    Returns
-    -------
-    out: torch.Tensor
-        Output tensor, shape ``(m, n)``.
-    """
-    _check_scale_granularity_mnk(scale_granularity_mnk)
-    _check_scale_major_mode_mxfp8(scale_major_mode)
-    if backend != "cute":
-        raise NotImplementedError(
-            f'Only backend="cute" is currently implemented; got backend="{backend}"'
-        )
-
-    if out_dtype is None:
-        out_dtype = out.dtype if out is not None else torch.bfloat16
-    if out_dtype != torch.bfloat16:
-        raise NotImplementedError(
-            f"Only out_dtype=torch.bfloat16 is supported; got {out_dtype}"
-        )
-
-    m, _ = a.shape
-    n = b.shape[0]
-    if out is None:
-        out = torch.empty((m, n), dtype=out_dtype, device=a.device)
-
-    get_gemm_sm120_module_cute_mxfp8().gemm_mxfp8_nt_groupwise(
-        a,
-        b,
-        a_scale,
-        b_scale,
-        out,
-        scale_granularity_mnk[0],
-        scale_granularity_mnk[1],
-        scale_granularity_mnk[2],
-    )
-    return out
-
-
-@supported_compute_capability([120, 121])
-@flashinfer_api
-def batch_gemm_mxfp8_nt_groupwise(
-    a: torch.Tensor,
-    b: torch.Tensor,
-    a_scale: torch.Tensor,
-    b_scale: torch.Tensor,
-    scale_granularity_mnk: Tuple[int, int, int] = (1, 1, 128),
-    scale_major_mode: Literal["MN"] = "MN",
-    lda: Optional[int] = None,
-    stride_a: Optional[int] = None,
-    ldb: Optional[int] = None,
-    stride_b: Optional[int] = None,
-    ldd: Optional[int] = None,
-    stride_d: Optional[int] = None,
-    backend: Literal["cute"] = "cute",
-    out: Optional[torch.Tensor] = None,
-    out_dtype: Optional[torch.dtype] = None,
-) -> torch.Tensor:
-    r"""Perform batched MXFP8 matmul with groupwise UE8M0 scaling (dense Batched entry).
-    Currently only supported on NVIDIA RTX PRO 6000 Blackwell (SM120) architecture.
-
-    Each batch ``a[i]`` is multiplied with ``b[i]``.
-
-    Parameters
-    ----------
-    a: torch.Tensor
-        Input tensor shape ``(num_groups, m, k)``, k-major (last stride must be 1),
-        data type is ``torch.float8_e4m3fn``.
-
-    b: torch.Tensor
-        Weight tensor shape ``(num_groups, n, k)``, k-major (last stride must be 1),
-        data type is ``torch.float8_e4m3fn``.
-
-    a_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``a``, shape ``(num_groups, m, k_align)``.
-        Data type ``torch.int32``.
-
-    b_scale: torch.Tensor
-        Int32-packed UE8M0 scale tensor for ``b``, shape ``(num_groups, n, k_align)``.
-        Data type ``torch.int32``.
-
-    scale_granularity_mnk: Tuple[int, int, int]
-        Accepted values: ``(1, 1, 128)`` (default) or ``(1, 1, 32)``.
-
-    lda, stride_a, ldb, stride_b, ldd, stride_d: Optional[int]
-        Optional explicit strides passed to the 6KD runner. If ``None``, defaults are
-        read from ``a.strides()`` / ``b.strides()`` / ``out.strides()`` (row stride for
-        ``ld*``, batch stride for ``stride_*``). Explicit pass overrides.
-
-    out: Optional[torch.Tensor]
-        Output tensor, shape ``(num_groups, m, n)``. If not specified, created.
-
-    out_dtype: Optional[torch.dtype]
-        Output dtype. Currently only ``torch.bfloat16`` is supported.
-
-    Returns
-    -------
-    out: torch.Tensor
-        Output tensor, shape ``(num_groups, m, n)``.
-    """
-    _check_scale_granularity_mnk(scale_granularity_mnk)
-    _check_scale_major_mode_mxfp8(scale_major_mode)
-    if backend != "cute":
-        raise NotImplementedError(
-            f'Only backend="cute" is currently implemented; got backend="{backend}"'
-        )
-
-    if out_dtype is None:
-        out_dtype = out.dtype if out is not None else torch.bfloat16
-    if out_dtype != torch.bfloat16:
-        raise NotImplementedError(
-            f"Only out_dtype=torch.bfloat16 is supported; got {out_dtype}"
-        )
-
-    num_groups, m, _ = a.shape
-    n = b.shape[1]
-    if out is None:
-        out = torch.empty((num_groups, m, n), dtype=out_dtype, device=a.device)
-
-    _lda = a.stride(1) if lda is None else lda
-    _stride_a = a.stride(0) if stride_a is None else stride_a
-    _ldb = b.stride(1) if ldb is None else ldb
-    _stride_b = b.stride(0) if stride_b is None else stride_b
-    _ldd = out.stride(1) if ldd is None else ldd
-    _stride_d = out.stride(0) if stride_d is None else stride_d
-
-    get_gemm_sm120_module_cute_mxfp8().batch_gemm_mxfp8_nt_groupwise(
-        a,
-        b,
-        a_scale,
-        b_scale,
-        out,
-        _lda,
-        _stride_a,
-        _ldb,
-        _stride_b,
-        _ldd,
-        _stride_d,
-        scale_granularity_mnk[0],
-        scale_granularity_mnk[1],
-        scale_granularity_mnk[2],
-    )
     return out
