+++ flashinfer/jit/gemm/__init__.py
@@ -26,7 +26,6 @@ from .core import (
     gen_mm_bf16_cublaslt_module,
     gen_gemm_sm100_module,
     gen_gemm_sm120_module,
-    gen_gemm_sm120_module_cute_mxfp8,
     gen_trtllm_gen_gemm_module,
     gen_trtllm_low_latency_gemm_module,
     gen_tgv_gemm_sm10x_module,
@@ -47,7 +46,6 @@ __all__ = [
     "gen_mm_bf16_cublaslt_module",
     "gen_gemm_sm100_module",
     "gen_gemm_sm120_module",
-    "gen_gemm_sm120_module_cute_mxfp8",
     "gen_trtllm_gen_gemm_module",
     "gen_trtllm_low_latency_gemm_module",
     "gen_tgv_gemm_sm10x_module",
